Deployment (#7)
* Restructures repo

* Updates filestructure

* Configurations for deployment

* Readds dashboard direct
