+++ .gitignore
@@ -1,3 +1,3 @@
 venv/
 __pycache__
-
+node_modules
