--- cytybyke/__init__.py
@@ -1,23 +1,0 @@
-from flask import Flask, render_template
-
-app = Flask(__name__)
-
-@app.route("/")
-def homepage():
-    return render_template("main.html")
-
-@app.route("/sign-in")
-def sign_in():
-    return render_template("sign-in.html")
-
-@app.route("/sign-up")
-def sign_up():
-    return render_template("sign-up.html")
-
-@app.route("/dashboard")
-def dashboard():
-    return render_template("dashboard.html")
-
-
-if __name__ == "__main__":
-    app.run(debug=True)
