+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "cytybyke",
+  "version": "1.0.0",
+  "description": "CytyByke is a web app for easily getting information about conditions at your favorite Citi Bike docks. Be confident that there will be a bike waiting for you at a station.",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/clyman23/cytybyke.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/clyman23/cytybyke/issues"
+  },
+  "homepage": "https://github.com/clyman23/cytybyke#readme",
+  "devDependencies": {
+    "firebase-tools": "^10.5.0"
+  }
+}
