+++ server/src/templates/dashboard.html
