+++ server/src/templates/index.html
@@ -3,6 +3,8 @@
 {% block content %}
 <h1>Hello world!</h1>
 <h1>Welcome to CytyByke</h1>
+<!-- Placeholder to show it is dynamic -->
+<p>{{ context.server_time }}</p>
 <div>
     <a class="btn btn-outline-primary" href="/sign-in">Sign in</a>
 </div>
